# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Translation API",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
        "/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "User Registration",
                "description": "Register a new user and return a token.",
                "operationId": "5d4fa5d5607ed675ba93907e1fc94534",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "john@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User registered successfully."
                    },
                    "422": {
                        "description": "Validation error."
                    }
                }
            }
        },
        "/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "User Login",
                "description": "Authenticate user and return access token.",
                "operationId": "67d6d0fdd54fba132e1b6f55ca254183",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "john@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login successful."
                    },
                    "401": {
                        "description": "Invalid credentials."
                    }
                }
            }
        },
        "/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "User Logout",
                "description": "Revoke the user's authentication token.",
                "operationId": "28d48c13fca984400ac8de99f54ee13a",
                "responses": {
                    "200": {
                        "description": "User logged out successfully."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/translations": {
            "get": {
                "tags": [
                    "Translations"
                ],
                "summary": "Retrieve all translations",
                "description": "Get all translations.",
                "operationId": "0436b18d23fb2409c24e5792abb2baf1",
                "responses": {
                    "200": {
                        "description": "List of translations"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            },
            "post": {
                "tags": [
                    "Translations"
                ],
                "summary": "Create a new translation",
                "description": "Store a new translation.",
                "operationId": "51818ae686a9f08bb44903ebb1eb616c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "locale",
                                    "key",
                                    "value"
                                ],
                                "properties": {
                                    "locale": {
                                        "type": "string",
                                        "example": "en"
                                    },
                                    "key": {
                                        "type": "string",
                                        "example": "greeting"
                                    },
                                    "value": {
                                        "type": "string",
                                        "example": "Hello"
                                    },
                                    "tags": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Translation created successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/translations/{id}": {
            "put": {
                "tags": [
                    "Translations"
                ],
                "summary": "Update a translation",
                "description": "Update an existing translation.",
                "operationId": "fb5c450fb84f94732d60db9d5c705020",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Translation ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "locale",
                                    "key",
                                    "value"
                                ],
                                "properties": {
                                    "locale": {
                                        "type": "string",
                                        "example": "fr"
                                    },
                                    "key": {
                                        "type": "string",
                                        "example": "welcome"
                                    },
                                    "value": {
                                        "type": "string",
                                        "example": "Bienvenue"
                                    },
                                    "tags": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Translation updated successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "404": {
                        "description": "Translation not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Translations"
                ],
                "summary": "Delete a translation",
                "description": "Delete a translation.",
                "operationId": "5fb2d061ac762da342efdf1414f3b6e4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Translation ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Translation deleted successfully"
                    },
                    "404": {
                        "description": "Translation not found"
                    }
                }
            }
        },
        "/translations/search": {
            "get": {
                "tags": [
                    "Translations"
                ],
                "summary": "Search translations",
                "description": "Search translations.",
                "operationId": "9b2805cd65f06c0bb36310c3e5035c28",
                "parameters": [
                    {
                        "name": "key",
                        "in": "query",
                        "description": "Search by key",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "value",
                        "in": "query",
                        "description": "Search by value",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "tags",
                        "in": "query",
                        "description": "Search by tags",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful search results"
                    },
                    "400": {
                        "description": "No search parameters provided"
                    }
                }
            }
        },
        "/translations/export": {
            "get": {
                "tags": [
                    "Translations"
                ],
                "summary": "Export translations",
                "description": "Export translations.",
                "operationId": "5303b390085c906651a02b91a31b5b8f",
                "responses": {
                    "200": {
                        "description": "Exported translations data"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Translations",
            "description": "API Endpoints for managing translations"
        },
        {
            "name": "Authentication",
            "description": "Authentication"
        }
    ]
}